quickstarts/speech-to-text/mic_stream_sample/requirements.txt
cffi==1.16.0
grpcio==1.62.1
grpcio-tools==1.62.1
numpy==1.26.4
protobuf==4.25.3
pycparser==2.22
python-dotenv==1.0.1
six==1.16.0
sounddevice==0.4.6
